step17 예제 추가

+++ Step17_JDBC/src/test/frame/MemberFrame.java
@@ -54,21 +54,22 @@ public class MemberFrame extends JFrame implements ActionListener, PropertyChang
       //페널째로 프레임의 북쪽에 배치 
       add(panel, BorderLayout.NORTH);      
       
-      //표형식으로 정보를 출력하기 위한 JTable
+      //표 형식으로 정보를 출력하기 위한 JTable
       table=new JTable();
       //칼럼명을 String[] 에 순서대로 준비
       String[] colNames= {"번호", "이름", "주소"};
       //테이블에 연결할 모델객체(테이블에 출력할 데이터를 가지고 있는 객체)
       model=new DefaultTableModel(colNames, 0) {
+    	  
     	  @Override
-    	public boolean isCellEditable(int row, int column) {
-    		// 0 번째 컬럼은 수정 불가능 하도록 false를 리턴하고
-    		if(column == 0) {
-    			return false;
-    		}else { // 나머지 컬럼은 수정 가능하도록 true를 리턴함.
-    			return true;
-    		}
-    	}
+    	  public boolean isCellEditable(int row, int column) {
+    		  // 0 번째 컬럼은 수정 불가능 하도록 false를 리턴하고
+    		  if(column == 0) {
+    			  return false;
+    		  }else { // 나머지 컬럼은 수정 가능하도록 true를 리턴함.
+    			  return true;
+    		  }
+    	  }
       }; 
       
       //모델을 테이블에 연결한다.
@@ -148,32 +149,38 @@ public class MemberFrame extends JFrame implements ActionListener, PropertyChang
         	 JOptionPane.showMessageDialog(this, "삭제할 row를 선택하세요!");
         	 return;
          }
-         //2. DefaultTableModel 에서 해당 인덱스의 table row 에서 삭제할 회원의 번호를 읽어와서(model 활용)
-         int num = (int) model.getValueAt(rowIndex, 0);
-         //3. MemberDao 객체를 이용해서 DB 에서 삭제하고
-         new MemberDao().delete(num);
-         //4. 새로 고침하기 
-         displayMember();
+         
+         int result = JOptionPane.showConfirmDialog(this, "선택된 row를 삭제하시겠습니까?");
+         // 만일 예를 눌렀다면
+         if(result == JOptionPane.YES_OPTION) {
+        	//2. DefaultTableModel 에서 해당 인덱스의 table row 에서 삭제할 회원의 번호를 읽어와서(model 활용)
+             int num = (int) model.getValueAt(rowIndex, 0);
+             //3. MemberDao 객체를 이용해서 DB 에서 삭제하고
+             new MemberDao().delete(num);
+             //4. 새로 고침하기 
+             displayMember();
+         }
       }
-      
    }
 
    // table에 특정 이벤트가 발생했을 때 호출되는 메소드
 	@Override
 	public void propertyChange(PropertyChangeEvent evt) {
-		if(!table.isEditing()) {
-			int rowIndex = table.getSelectedRow();
-			if(rowIndex != -1) {
-				int num = (int) model.getValueAt(rowIndex, 0);
-				 String name = (String)model.getValueAt(rowIndex, 1);
-		         String addr = (String)model.getValueAt(rowIndex, 2);
-		         MemberDto dto = new MemberDto();
-		         dto.setName(name);
-		         dto.setAddr(addr);
-		         dto.setNum(num);
-		         new MemberDao().update(dto);
-		         return;
-	         }
+		// 만일 테이블의 수정 사항을 DB에 수정 반영할 적당한 시점이 되면
+		if(evt.getPropertyName().equals("tableCellEditor") && !table.isEditing()) {
+			// 현재 선택된 row의 정보를 DB에 수정 반영함.
+			// 변환된 값을 읽어와서 DB에 반영함.
+			// 수정된 컬럼에 있는 row 전체의 값을 읽어옴.
+			int selectedIndex = table.getSelectedRow();
+			int num = (int)model.getValueAt(selectedIndex, 0);
+			 String name = (String)model.getValueAt(selectedIndex, 1);
+	         String addr = (String)model.getValueAt(selectedIndex, 2);
+	         // 수정할 회원의 정보를 MemberDto 객체에 담고
+	         MemberDto dto = new MemberDto(num, name, addr);
+	         // DB에 저장하기
+	         new MemberDao().update(dto);
+	         // 선택된 row clear
+	         table.clearSelection();
 		}
 	}
 }
